RF: Move ReadOnlyDict into a new datalad/support/collections.py
this construct will be useful in other places as well
